Ticket send and edit works
the ticket in the list of users can be edited an saved in the database. Only tickets that are not closed are seen in the listview.  listview has  double klik mouse event to open the ticket. little bug are out of the applications Except filling combox of the ticket itself. So open close and Pending